Add formatting settings
