--- scripts/helpers/nullability.cs
@@ -1,21 +1,0 @@
-using Godot;
-using System;
-using System.Diagnostics.CodeAnalysis;
-
-public static class NullHelpers
-{
-    public static T NotNull<T>([NotNull] this T? obj, string objName)
-    {
-    	return obj ?? throw new ArgumentNullException(objName);
-    }
-
-	public static T GDNotNull<T>([NotNull] this T? node, string nodeName) where T : GodotObject
-	{
-		if (node == null || !GodotObject.IsInstanceValid(node))
-		{
-			throw new ArgumentNullException(nodeName);
-		}
-
-		return node;
-	}
-}
