+++ scripts/logic/GameManager.cs
@@ -1,37 +1,37 @@
-using Godot;
 using System.Collections.Generic;
+using Godot;
 
 public partial class GameManager : Node
 {
-	public const string CurrentSeedEntry = "currentSeed";
+    public const string CurrentSeedEntry = "currentSeed";
 
-	private PackedScene? _sceneTempLevel;
-	private MainMenu? _mainMenu;
+    private PackedScene? _sceneTempLevel;
+    private MainMenu? _mainMenu;
 
-	private readonly Dictionary<string, int> _gameInfo = new Dictionary<string, int> { { CurrentSeedEntry, 0 } };
-	
-	public override void _Ready()
-	{
-		_sceneTempLevel = GD.Load<PackedScene>("uid://chvfto1w6w3um").NotNull(nameof(_sceneTempLevel));
-		_mainMenu = GetNode<MainMenu>("Main Menu").NotNull(nameof(_mainMenu));
+    private readonly Dictionary<string, int> _gameInfo = new Dictionary<string, int> { { CurrentSeedEntry, 0 } };
 
-		_mainMenu.SeedSet += OnMainMenuSeedSet;
-		_mainMenu.PlayPressed += OnMainMenuPlayPressed;
-	}
+    public override void _Ready()
+    {
+        _sceneTempLevel = GD.Load<PackedScene>("uid://chvfto1w6w3um").NotNull(nameof(_sceneTempLevel));
+        _mainMenu = GetNode<MainMenu>("Main Menu").NotNull(nameof(_mainMenu));
 
-	private void OnMainMenuPlayPressed()
-	{
-		_sceneTempLevel.NotNull(nameof(_sceneTempLevel));
-		_mainMenu.NotNull(nameof(_mainMenu));
-	
-		Node tempLevel = _sceneTempLevel.Instantiate();
-		GetTree().Root.AddChild(tempLevel);
+        _mainMenu.SeedSet += OnMainMenuSeedSet;
+        _mainMenu.PlayPressed += OnMainMenuPlayPressed;
+    }
 
-		_mainMenu.QueueFree();
-	}
+    private void OnMainMenuPlayPressed()
+    {
+        _sceneTempLevel.NotNull(nameof(_sceneTempLevel));
+        _mainMenu.NotNull(nameof(_mainMenu));
 
-	private void OnMainMenuSeedSet(int newSeed)
-	{
-		_gameInfo[CurrentSeedEntry] = newSeed;
-	}
+        Node tempLevel = _sceneTempLevel.Instantiate();
+        GetTree().Root.AddChild(tempLevel);
+
+        _mainMenu.QueueFree();
+    }
+
+    private void OnMainMenuSeedSet(int newSeed)
+    {
+        _gameInfo[CurrentSeedEntry] = newSeed;
+    }
 }
