+++ scripts/logic/Health.cs
@@ -1,29 +1,29 @@
-using Godot;
 using System.Diagnostics;
+using Godot;
 
 public partial class Health : Node
 {
-	[Export] public int MaxHealth;
+    [Export] public int MaxHealth;
 
-	[Signal] public delegate void HealthEmptyEventHandler();
-	[Signal] public delegate void HealthPoolChangedEventHandler();
+    [Signal] public delegate void HealthEmptyEventHandler();
+    [Signal] public delegate void HealthPoolChangedEventHandler();
 
-	public int HealthPool;
-	
-	public override void _Ready()
-	{
-		Debug.Assert(MaxHealth > 0);
-		HealthPool = MaxHealth;
-	}
+    public int HealthPool;
 
-	public void ChangeHealth(int value)
-	{
-		HealthPool = Mathf.Clamp(HealthPool + value, 0, MaxHealth);
-		if (HealthPool == 0)
-		{
-			EmitSignal(SignalName.HealthEmpty);
-		}
+    public override void _Ready()
+    {
+        Debug.Assert(MaxHealth > 0);
+        HealthPool = MaxHealth;
+    }
 
-		EmitSignal(SignalName.HealthPoolChanged);
-	}
+    public void ChangeHealth(int value)
+    {
+        HealthPool = Mathf.Clamp(HealthPool + value, 0, MaxHealth);
+        if (HealthPool == 0)
+        {
+            EmitSignal(SignalName.HealthEmpty);
+        }
+
+        EmitSignal(SignalName.HealthPoolChanged);
+    }
 }
