+++ scripts/logic/HealthBar.cs
@@ -2,22 +2,22 @@ using Godot;
 
 public partial class HealthBar : ProgressBar
 {
-	[Export] private Health? _healthPool;
-	
-	public override void _Ready()
-	{
-		_healthPool.NotNull(nameof(_healthPool));
+    [Export] private Health? _healthPool;
 
-		_healthPool.HealthPoolChanged += OnHealthPoolChanged;
+    public override void _Ready()
+    {
+        _healthPool.NotNull(nameof(_healthPool));
 
-		MaxValue = _healthPool.MaxHealth;
-		Value = _healthPool.HealthPool;
-	}
+        _healthPool.HealthPoolChanged += OnHealthPoolChanged;
 
-	private void OnHealthPoolChanged()
-	{
-		_healthPool.NotNull(nameof(_healthPool));
+        MaxValue = _healthPool.MaxHealth;
+        Value = _healthPool.HealthPool;
+    }
 
-		Value = _healthPool.HealthPool;
-	}
+    private void OnHealthPoolChanged()
+    {
+        _healthPool.NotNull(nameof(_healthPool));
+
+        Value = _healthPool.HealthPool;
+    }
 }
