+++ scripts/player/Arm.cs
@@ -2,8 +2,8 @@ using Godot;
 
 public partial class Arm : Node2D
 {
-	public override void _Process(double delta)
-	{
-		LookAt(GetGlobalMousePosition());
-	}
+    public override void _Process(double delta)
+    {
+        LookAt(GetGlobalMousePosition());
+    }
 }
